Move AllyariaColor files into Palette folder/namespace to create separation with future Theming components.
