+++ src/Allyaria.Theming/Palette/AllyariaColor.Fields.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.RegularExpressions;
 
-namespace Allyaria.Theming;
+namespace Allyaria.Theming.Palette;
 
 public readonly partial struct AllyariaColor
 {
