+++ src/Allyaria.Theming/Palette/AllyariaColor.Operators.cs
@@ -1,4 +1,4 @@
-﻿namespace Allyaria.Theming;
+﻿namespace Allyaria.Theming.Palette;
 
 public readonly partial struct AllyariaColor
 {
